requirements.txt
agentscope
python-dotenv
PyMuPDF
pillow
opencv-python
openai-whisper